[*] Introduce TaskSet and use it instead of std::set.
Profiling showed that a lot of time was wasted in sorting
a binary tree (which is what std::set is). It only was used
to store a collection of unique task IDs. Since we know that
there's a fixed ammount of task IDs and also that they all
start from 0 and end with MAX_TASKS-1, we can just use a
bit set to store the state of the tasks.

So we don't have to maintain a cache-unfriendly binary tree
for the sake of checking if some task was or wasn't finished.

The result: ~20% win in performance.