[*] Use separated task queues for each worker.
Profiling showed that a lot of time is wasted in syncronization
primitives. The problem is that all threads share task queue and
finihed task list. So we have a terrible serialization in two hot
spots in a compartment with syncronization overhead.

Now I made a separated task queue for each thread. So the workers
don't have to fight each other to get the next task to exeute.

Result: ~53% win in performance.