[*] Don't do apriori-false-resulting request.
Al lot of time is still wasted in synchronization primitives. One
of possible problems is that the blocking request to set of
finished tasks is called highly actively.

If a task wasn't scheduled yet then we have no reason to check if
it's finished, so slightly modified code to use this fact.

Result: ~20% win in performance.