[*] Switch to lock-free reversed build graph.
Profiling shows that the most time-consuming part of program
runtime is still synchronization. It was discovered that under
certain conditions we can use lock-free data structure to aquire
next task to execute, but this leads to one downside: one target
can't be required by several targets.

The test dataset passes the requirement so the solution showed an
unexpected speed-up.

Result: ~1000% win in performance.