[*] Make task the lock-free queue thread-local.
According to the profile we have contention again: most of time was
wasted in aquire_next_independent_task function which traversed the
shared task queue. It's not as bad as lock contention, but bad
enough to be fixed.

But I'm concerned about how much time is reversed build graph
construction taking.

Also I think the task scheduling can be cache-unfriendly speaking
of the task-to-dependers mapping.

Result: %8 - %39 win in performance (it's just more stable now).